[add]pay-function
