+++ app/controllers/cards_controller.rb
@@ -63,6 +63,6 @@ class CardsController < ApplicationController
   end
 
   def set_card
-    @card = Card.find_by(user_id: current_user.id)
+    @card = Card.where(user_id: current_user.id).first
   end
 end
