+++ app/controllers/items_controller.rb
@@ -1,5 +1,11 @@
 class ItemsController < ApplicationController
-  before_action :set_item, except: %i[index new create]
+  require 'payjp'
+
+  before_action :set_item, except: %i[index new create purchaseCompleted]
+  before_action :set_card, only: %i[purchaseConfilmation pay]
+  before_action :set_sending_destinations, only: %i[purchaseConfilmation] 
+  before_action :set_api_key
+  
 
   def index
     @items = Item.all
@@ -39,6 +45,30 @@ class ItemsController < ApplicationController
     end
   end
 
+  def purchaseConfilmation
+    if @card.blank?
+      flash[:alert] = '購入前にクレジットカードを登録してください'
+      redirect_to controller: "cards", action: "new"
+    else
+      set_item
+      set_sending_destinations
+      set_customer
+      set_card_information
+    end
+  end
+
+  def pay
+    charge = Payjp::Charge.create(
+      amount: @item.price, #支払金額を引っ張ってくる
+      customer: @card.customer_id,  #顧客ID
+      currency: 'jpy',              #日本円
+    )
+    redirect_to purchaseCompleted_item_path #購入完了ページへ
+  end
+
+  def purchaseCompleted
+  end
+
   private
 
   def item_params
@@ -48,4 +78,25 @@ class ItemsController < ApplicationController
   def set_item
     @item = Item.find(params[:id])
   end
+
+  def set_api_key
+    Payjp.api_key = Rails.application.credentials[:payjp][:secret_key]
+  end
+
+  def set_customer # 保管した顧客IDでpayjpから情報取得
+    @customer = Payjp::Customer.retrieve(@card.customer_id)
+  end
+
+  def set_card_information # 保管したカードIDでpayjpから情報取得、カード情報表示のためインスタンス変数に代入
+    @card_information = @customer.cards.retrieve(@card.card_id)
+  end
+
+  def set_card
+    @card = Card.where(user_id: current_user.id).first
+  end
+
+  def set_sending_destinations
+    @address = SendingDestination.where(user_id: current_user.id).first
+  end
+
 end
