+++ app/models/card.rb
@@ -1,4 +1,3 @@
 class Card < ApplicationRecord
   belongs_to :user
-  belongs_to :purchase
 end
