+++ app/models/item.rb
@@ -3,7 +3,6 @@ class Item < ApplicationRecord
   # belongs_to :brand
   has_many :images, dependent: :destroy
   accepts_nested_attributes_for :images, allow_destroy: true
-  belongs_to :purchase
   # has_one :user_evaluation
 
   # belongs_to_active_hash :size
