--- app/models/purchase.rb
@@ -1,4 +1,0 @@
-class Purchase < ApplicationRecord
-  belongs_to :item
-  belongs_to :card
-end
