+++ app/views/items/purchaseCompleted.html.haml
@@ -1,0 +1,5 @@
+.app-name
+  =link_to image_tag("logo.png",size:"200x60"),"root_path"
+
+%h1 購入完了しました
+= link_to "トップページへ戻る", root_path
