+++ app/views/items/purchaseConfilmation.html.haml
@@ -1,0 +1,95 @@
+.app-name
+  =link_to image_tag("logo.png",size:"200x60"),"root_path"
+%buy-main
+  = form_tag(action: :pay, method: :post) do
+    %h1 購入内容の確認
+    = image_tag @item.images[0].src.url, class:""
+    %section
+      %p 商品名
+      = @item.name
+      %p 
+        = number_to_currency(@item.price, format: "%u%n", unit: "¥", precision: 0)
+        %span
+          (税込)送料込み
+      %section
+        %p 支払い金額
+        = number_to_currency(@item.price, format: "%u%n", unit:"¥", precision: 0)
+        %section
+          %h3 支払方法
+          .cards__list
+            %figure
+              -# = image_tag "#{@card_src}",alt: @card_brand, id: "card_image"
+            .cards__number
+              = "**** **** **** " + @card_information.last4
+            .cards__number
+              - exp_month = @card_information.exp_month.to_s
+              - exp_year = @card_information.exp_year.to_s.slice(2,3)
+              = exp_month + " / " + exp_year
+        %section
+          %h3 配送先
+          .user--info
+            = "〒" + @address.post_code.to_s.insert(3, "-")
+            %br
+            = @address.prefecture_code
+            %br
+            = @address.city
+            = @address.house_number
+            = @address.building_name
+            %br
+            = @address.destination_family_name
+            = @address.destination_first_name
+          %section
+            = submit_tag("購入する", class: "pay")
+
+///後で消す
+-# %main.buy-main 
+-#   = form_tag(action: :pay, method: :post) do
+-#     .buy-item-container
+-#       %h2.buy-item-head 購入内容の確認
+-#       %section.buy-content.buy-item
+-#         .buy-content-inner
+-#           .buy-item-box
+-#             .buy-item-image
+-#               =image_tag @item.images[0].src.url, class:"buy-image"
+-#             .buy-item-detail
+-#               %p.buy-item-name
+-#                 =@item.name
+-#                 %p.buy-item-price.bold
+-#                   = number_to_currency(@item.price, format: "%u%n",unit:"¥", precision: 0)
+-#                   %span.item-shipping-fee.f14.bold
+-#                     (税込)送料込み
+-#       %section.buy-content.buy-user-info
+-#         .buy-content-inner
+-#           %ul.buy-price-table
+-#             %li.buy-price-row.buy-you-pay.bold
+-#             %li.buy-price-cell 支払い金額
+-#             %li.buy-price-cell
+-#               = number_to_currency(@item.price,format: "%u%n",unit:"¥",precision: 0)
+-#           %section.buy-content.buy-user-info
+-#         .buy-content-inner
+-#           %h3 支払方法
+-#           .payment-content__cards__list
+-#             %figure
+-#               -# = image_tag "#{@card_src}",alt: @card_brand, id: "card_image"
+-#             .payment-content__cards__list__number
+-#               = "**** **** **** " + @card_information.last4
+-#             .payment-content__cards__list__number
+-#               - exp_month = @card_information.exp_month.to_s
+-#               - exp_year = @card_information.exp_year.to_s.slice(2,3)
+-#               = exp_month + " / " + exp_year
+-#           %section.buy-content.buy-user-info
+-#         .buy-content-inner
+-#           %h3 配送先
+-#           %address.buy-user-info-text
+-#             = @address.post_code
+-#             %br
+-#             = @address.prefecture_code
+-#             %br
+-#             = @address.city
+-#             = @address.house_number
+-#             = @address.building_name
+-#             %br
+-#             = @address.destination_family_name
+-#             = @address.destination_first_name
+-#           %section.buy-content.buy-user-info
+-#           = submit_tag("購入する", class:"pay")
