--- app/views/purchases/done.html.haml
@@ -1,2 +1,0 @@
-%h1 Purchases#done
-%p Find me in app/views/purchases/done.html.haml
