--- app/views/purchases/index.html.haml
@@ -1,2 +1,0 @@
-%h1 Purchases#index
-%p Find me in app/views/purchases/index.html.haml
