--- db/migrate/20200601043442_create_purchases.rb
@@ -1,8 +1,0 @@
-class CreatePurchases < ActiveRecord::Migration[5.2]
-  def change
-    create_table :purchases do |t|
-
-      t.timestamps
-    end
-  end
-end
