--- spec/factories/purchases.rb
@@ -1,5 +1,0 @@
-FactoryBot.define do
-  factory :purchase do
-    
-  end
-end
