--- spec/requests/purchases_request_spec.rb
@@ -1,19 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Purchases", type: :request do
-
-  describe "GET /index" do
-    it "returns http success" do
-      get "/purchases/index"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /done" do
-    it "returns http success" do
-      get "/purchases/done"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-end
